+++ global_modules.js
@@ -2,7 +2,6 @@ module.exports = {
 	EventEmitter: "events",
 	FS: "fs-extra",
 	Path: "path",
-	Seed: "seed",
 	URL: "url",
 	URLJoin: "url-join",
 	Util: "util"
