+++ index.js
@@ -1,6 +1,13 @@
-var Seed = function() {
-	return require("./Core");
-};
+var globalModules = require("./global_modules");
+
+for(var module in globalModules) {
+	global[module] = require(globalModules[module]);
+}
+
+// Core
+var Seed = require("./Core");
+
+global.Seed = Seed;
 
 // Abstract
 Seed.Component = require("./Abstract/Component");
@@ -11,4 +18,17 @@ Seed.LessHandler = require("./LessHandler");
 Seed.Server = require("./Server");
 Seed.TemplateEngine = require("./TemplateEngine");
 
-module.exports = Seed;
+global.dumpError = function(err) {
+	if (typeof err === 'string')
+		console.error(err);
+	else if (typeof err === 'object') {
+		if (err.message) {
+			console.error('\nMessage: ' + err.message)
+		}
+		if (err.stack) {
+			console.error('\nStacktrace:')
+			console.error('====================')
+			console.error(err.stack);
+		}
+	}
+}
